GH-86: Add `Codec` transformers factories
Fixes GH-86 (https://github.com/spring-projects/spring-integration-java-dsl/issues/86)

* Add `Transformers` `encoding()` and `decoding()` factories for the `EncodingPayloadTransformer` and `DecodingTransformer` respectively
* Refactor `IntegrationFlowTests`  a bit to extract `enrich()` tests into a new `TransformerTests`
* Add `encoding()/decoding()` tests
